# backend/openapi.yaml
openapi: 3.1.0
info:
  title: "KAVACH \u2014 AI-Powered DevSecOps Platform"
  description: "KAVACH scans banking application repositories for vulnerabilities,\
    \ calculates a Banking Risk Score (BRS), maps findings to regulatory controls\
    \ (RBI / PCI-DSS / SWIFT), and generates AI-powered audit reports.\n\n## Pipeline\n\
    `GitHub webhook` \u2192 `Create Scan Job` \u2192 `Queue` \u2192 `Distributed workers`\
    \ \u2192 `Aggregation` \u2192 `Risk Engine` \u2192 `Compliance Engine` \u2192\
    \ `AI Explanation` \u2192 `Reports` \u2192 `Dashboard` \u2192 `Notifications`\
    \ \u2192 `Archive`\n\nA scan enters this pipeline either via `POST /scan/repository`\
    \ / `POST /scan` (an authenticated user submitting a URL or archive) or `POST\
    \ /webhooks/github` (an unauthenticated, HMAC-verified push event) \u2014 both\
    \ converge on the same queue \u2192 distributed-worker \u2192 aggregation path,\
    \ so everything downstream behaves identically regardless of how the scan was\
    \ triggered."
  contact:
    name: KAVACH Platform Team
    url: https://github.com/kavach-platform/kavach
  license:
    name: "Proprietary \u2014 All Rights Reserved"
  version: 1.0.0
paths:
...
  /api/v1/risk/modules:
    get:
      tags:
      - Risk Configuration
      summary: List Business Modules
      description: "List every configured business module \u2014 the DB-driven replacement\
        \ for the old hardcoded MODULE_WEIGHTS."
      operationId: list_business_modules_api_v1_risk_modules_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/BusinessModuleResponse'
                type: array
                title: Response List Business Modules Api V1 Risk Modules Get
      security:
      - OAuth2PasswordBearer: []
    post:
      tags:
      - Risk Configuration
      summary: Create Business Module
      description: "Add a new business module \u2014 this is what 'fully extensible'\
        \ means in practice: no code change, no deploy."
      operationId: create_business_module_api_v1_risk_modules_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessModuleCreateRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessModuleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/risk/modules/{module_name}:
    patch:
      tags:
      - Risk Configuration
      summary: Update Business Module
      description: "Adjust a module's weights/keywords \u2014 takes effect on the\
        \ very next scan, no deploy."
      operationId: update_business_module_api_v1_risk_modules__module_name__patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: module_name
        in: path
        required: true
        schema:
          type: string
          title: Module Name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessModuleUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessModuleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Risk Configuration
      summary: Delete Business Module
      description: "Remove a module. The fallback ('General', is_default=True) can't\
        \ be deleted \u2014 scoring needs\nsomewhere to land a finding that matches\
        \ no other module's keywords."
      operationId: delete_business_module_api_v1_risk_modules__module_name__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: module_name
        in: path
        required: true
        schema:
          type: string
          title: Module Name
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/risk/factor-weights:
    get:
      tags:
      - Risk Configuration
      summary: List Risk Factor Weights
      description: List every factor's weight in the BRS blend (CVSS, exploitability,
        business criticality, ...).
      operationId: list_risk_factor_weights_api_v1_risk_factor_weights_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/RiskFactorWeightResponse'
                type: array
                title: Response List Risk Factor Weights Api V1 Risk Factor Weights
                  Get
      security:
      - OAuth2PasswordBearer: []
  /api/v1/risk/factor-weights/{factor_name}:
    patch:
      tags:
      - Risk Configuration
      summary: Update Risk Factor Weight
      description: "Set a factor's weight \u2014 0 disables it entirely (the blend\
        \ normalizes\nby the sum of active weights, so this is safe). Upserts: naming\
        \ a\nfactor that doesn't exist yet creates it, which is how a newly\ncode-added\
        \ factor (see brs_engine.py's registry note) gets its\ndefault weight without\
        \ a migration."
      operationId: update_risk_factor_weight_api_v1_risk_factor_weights__factor_name__patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: factor_name
        in: path
        required: true
        schema:
          type: string
          title: Factor Name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiskFactorWeightUpsertRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskFactorWeightResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/risk/preview:
    post:
      tags:
      - Risk Configuration
      summary: Preview Score
      description: "Score a synthetic finding against the *current* database configuration\n\
        \u2014 the concrete way to see what changing a module's criticality_weight\n\
        or a factor's weight actually does before it affects a real scan.\nReuses\
        \ brs_engine's own DB-loaders (repo-backed, falling back to\ndefaults if unconfigured)\
        \ rather than duplicating that logic here."
      operationId: preview_score_api_v1_risk_preview_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScorePreviewRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScorePreviewResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
...
components:
  schemas:
...
    BusinessModuleCreateRequest:
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 1
          title: Name
        keywords:
          items:
            type: string
          type: array
          title: Keywords
        criticality_weight:
          type: number
          maximum: 10.0
          minimum: 0.0
          title: Criticality Weight
        asset_value:
          type: number
          maximum: 10.0
          minimum: 0.0
          title: Asset Value
        is_internet_facing_default:
          type: boolean
          title: Is Internet Facing Default
          default: false
        is_default:
          type: boolean
          title: Is Default
          default: false
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      type: object
      required:
      - name
      - criticality_weight
      - asset_value
      title: BusinessModuleCreateRequest
    BusinessModuleResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        keywords:
          items:
            type: string
          type: array
          title: Keywords
        criticality_weight:
          type: number
          title: Criticality Weight
        asset_value:
          type: number
          title: Asset Value
        is_internet_facing_default:
          type: boolean
          title: Is Internet Facing Default
        is_default:
          type: boolean
          title: Is Default
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      type: object
      required:
      - id
      - name
      - keywords
      - criticality_weight
      - asset_value
      - is_internet_facing_default
      - is_default
      title: BusinessModuleResponse
    BusinessModuleUpdateRequest:
      properties:
        keywords:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Keywords
        criticality_weight:
          anyOf:
          - type: number
            maximum: 10.0
            minimum: 0.0
          - type: 'null'
          title: Criticality Weight
        asset_value:
          anyOf:
          - type: number
            maximum: 10.0
            minimum: 0.0
          - type: 'null'
          title: Asset Value
        is_internet_facing_default:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Internet Facing Default
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      type: object
      title: BusinessModuleUpdateRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    RiskFactorWeightResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        factor_name:
          type: string
          title: Factor Name
        weight:
          type: number
          title: Weight
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      type: object
      required:
      - id
      - factor_name
      - weight
      title: RiskFactorWeightResponse
    RiskFactorWeightUpsertRequest:
      properties:
        weight:
          type: number
          minimum: 0.0
          title: Weight
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      type: object
      required:
      - weight
      title: RiskFactorWeightUpsertRequest
...
    ScorePreviewRequest:
      properties:
        title:
          type: string
          title: Title
          default: Preview finding
        severity:
          type: string
          pattern: ^(CRITICAL|HIGH|MEDIUM|LOW|INFO)$
          title: Severity
          default: HIGH
        category:
          type: string
          title: Category
          default: sql_injection
        cvss:
          type: number
          maximum: 10.0
          minimum: 0.0
          title: Cvss
          default: 7.5
        file_path:
          anyOf:
          - type: string
          - type: 'null'
          title: File Path
        description:
          type: string
          title: Description
          default: ''
        cve:
          anyOf:
          - type: string
          - type: 'null'
          title: Cve
        compliance_framework_count:
          type: integer
          maximum: 3.0
          minimum: 0.0
          title: Compliance Framework Count
          default: 0
        historical_incident_count:
          type: integer
          minimum: 0.0
          title: Historical Incident Count
          default: 0
      type: object
      title: ScorePreviewRequest
      description: "A synthetic finding to run through the current DB-configured\n\
        formula \u2014 lets an operator see exactly how a weight/module change\nwould\
        \ affect scoring before a real scan hits it."
    ScorePreviewResponse:
      properties:
        brs:
          type: number
          title: Brs
        module:
          type: string
          title: Module
        sub_scores:
          additionalProperties:
            type: number
          type: object
          title: Sub Scores
        factor_weights:
          additionalProperties:
            type: number
          type: object
          title: Factor Weights
      type: object
      required:
      - brs
      - module
      - sub_scores
      - factor_weights
      title: ScorePreviewResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login
...